Renamed Login to Registr and Token to Login Activities
