+++ app/src/main/AndroidManifest.xml
@@ -30,23 +30,23 @@
         </activity>
 
         <activity
-            android:name=".ui.activities.TokenActivity_"
+            android:name=".ui.activities.LoginActivity_"
             android:theme="@style/AppTheme.NoActionBar" >
             <meta-data android:name="android.support.PARENT_ACTIVITY"
                 android:value=".ui.activities.SplashActivity_"/>
         </activity>
 
         <activity
-            android:name=".ui.activities.LoginActivity_"
+            android:name=".ui.activities.RegistrActivity_"
             android:theme="@style/AppTheme.NoActionBar" >
             <meta-data android:name="android.support.PARENT_ACTIVITY"
-                android:value=".ui.activities.TokenActivity_"/>
+                android:value=".ui.activities.LoginActivity_"/>
         </activity>
 
         <activity android:name=".ui.activities.MainActivity_"
             android:theme="@style/AppTheme.NoActionBar" >
         <meta-data android:name="android.support.PARENT_ACTIVITY"
-            android:value=".ui.activities.LoginActivity_"/>
+            android:value=".ui.activities.RegistrActivity_"/>
         </activity>
 
         <activity android:name=".ui.activities.AddExpenseActivity_">
