+++ app/src/main/java/com/hardteam/moneytracker/ui/activities/LoginActivity.java
@@ -3,9 +3,6 @@ package com.hardteam.moneytracker.ui.activities;
 import android.content.Intent;
 import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.text.Layout;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -14,80 +11,87 @@ import com.hardteam.moneytracker.MoneyTrackerApplication;
 import com.hardteam.moneytracker.R;
 import com.hardteam.moneytracker.rest.RestService;
 import com.hardteam.moneytracker.rest.model.UserLoginModel;
-import com.hardteam.moneytracker.rest.model.UserRegistrationModel;
 import com.hardteam.moneytracker.util.Constants;
 import com.hardteam.moneytracker.util.NetworkStatusChecker;
 
-import org.androidannotations.annotations.AfterViews;
 import org.androidannotations.annotations.Background;
 import org.androidannotations.annotations.Click;
 import org.androidannotations.annotations.EActivity;
 import org.androidannotations.annotations.ViewById;
 
 /**
- * Created by RG on 25.12.2015.
+ * Created by rg on 12/29/15.
  */
 
 @EActivity(R.layout.activity_login)
-public class LoginActivity extends AppCompatActivity {
+public class LoginActivity extends AppCompatActivity{
 
-    private static final String LOG_VIEW = LoginActivity.class.getSimpleName();
+    @ViewById(R.id.login_field_token)
+    EditText loginFieldToken;
 
-    @ViewById(R.id.login_field)
-    EditText loginField;
+    @ViewById(R.id.password_field_token)
+    EditText passFieldToken;
 
-    @ViewById(R.id.password_field)
-    EditText passField;
+    @ViewById(R.id.token_button_get)
+    Button buttonToken;//Login
 
-    @ViewById(R.id.login_button_ok)
-    Button buttonOk;
+    @ViewById(R.id.token_activity)
+    View tokenActivityLayout;
 
-    @ViewById(R.id.login_activity_container)
-    View loginActivityLayout;
-
-    @Click(R.id.login_button_ok)
-    void loginOk()
+    @Click(R.id.token_button_get)
+    void buttonToken()//Login
     {
 
         if(NetworkStatusChecker.isNetworkAvailable(this))
         {
-            registerUser();
+            tokenIn();
         }
         else
         {
-            Snackbar.make(loginActivityLayout, Constants.noInternet, Snackbar.LENGTH_LONG).show();
+            Snackbar.make(tokenActivityLayout, Constants.noInternet, Snackbar.LENGTH_LONG).show();
         }
+    }
 
+    @Click(R.id.registration_token_text)
+    void textClickRegistration()
+    {
+        startRegistrationActivity();
     }
 
     @Background
-    public void registerUser()
+    void tokenIn()
     {
-        String loginUser = loginField.getText().toString();
-        String passwordUser = passField.getText().toString();
 
-        if( loginUser.length() < 5 || passwordUser.length() < 5 )
+        String loginUserToken = loginFieldToken.getText().toString();
+        String passwordUserToken = passFieldToken.getText().toString();
+
+        if( loginUserToken.length() < 5 || passwordUserToken.length() < 5 )
         {
-            Snackbar.make(loginActivityLayout, Constants.loginPassLength, Snackbar.LENGTH_LONG).show();
+            Snackbar.make(tokenActivityLayout, Constants.loginPassLength, Snackbar.LENGTH_LONG).show();
         }
 
         else
         {
             RestService restService = new RestService();
-            UserRegistrationModel userRegistrationModel = restService.register(loginUser, passwordUser);
+            UserLoginModel userLoginModel = restService.login(loginUserToken,passwordUserToken);
 
-            if(userRegistrationModel.getStatus().equalsIgnoreCase(Constants.success))
+            if(userLoginModel.getStatus().equalsIgnoreCase(Constants.success))
             {
-                UserLoginModel userLoginModel = restService.login(loginUser,passwordUser);
                 MoneyTrackerApplication.setAuthToken(userLoginModel.getAuthToken());
-
                 startMainActivity();
             }
-            else
+            else if(userLoginModel.getStatus().equalsIgnoreCase(Constants.wrongPassword))
             {
-                Snackbar.make(loginActivityLayout, Constants.registrationBusy, Snackbar.LENGTH_LONG).show();
+                Snackbar.make(tokenActivityLayout, Constants.wrongPassword, Snackbar.LENGTH_LONG).show();
             }
-
+            else if(userLoginModel.getStatus().equalsIgnoreCase(Constants.wrongLogin))
+            {
+                Snackbar.make(tokenActivityLayout, Constants.wrongLogin, Snackbar.LENGTH_LONG).show();
+            }
+            else if(userLoginModel.getStatus().equalsIgnoreCase(Constants.error))
+            {
+                Snackbar.make(tokenActivityLayout, Constants.error, Snackbar.LENGTH_LONG).show();
+            }
         }
     }
 
@@ -96,4 +100,12 @@ public class LoginActivity extends AppCompatActivity {
         Intent intentMainActivity = new Intent(this, MainActivity_.class);
         startActivity(intentMainActivity);
     }
+
+    public void startRegistrationActivity()
+    {
+        Intent intentRegistrationActivity = new Intent(this, RegistrActivity_.class);
+        startActivity(intentRegistrationActivity);
+    }
+
+
 }
