+++ app/src/main/res/layout/activity_login.xml
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
+
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_height="match_parent"
     android:layout_width="match_parent"
-    android:id="@+id/login_activity_container"
+    android:id="@+id/token_activity"
     android:background="@color/green">
 
     <LinearLayout
@@ -17,7 +18,7 @@
         <EditText
             android:hint="Login"
             android:paddingLeft="@dimen/padding_left_text"
-            android:id="@+id/login_field"
+            android:id="@+id/login_field_token"
             android:maxLength="10"
             android:layout_marginTop="@dimen/layout_margin_top"
             android:textSize="@dimen/text_size_login"
@@ -29,7 +30,7 @@
         <EditText
             android:hint="Password"
             android:paddingLeft="@dimen/padding_left_text"
-            android:id="@+id/password_field"
+            android:id="@+id/password_field_token"
             android:maxLength="10"
             android:layout_marginTop="@dimen/layout_margin_top"
             android:singleLine="true"
@@ -38,15 +39,25 @@
             android:background="@color/white"
             android:layout_width="match_parent"
             android:layout_height="@dimen/layout_height_text_field"/>
-
         <Button
-            android:id="@+id/login_button_ok"
+            android:id="@+id/token_button_get"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginTop="@dimen/padding_top"
             android:padding="@dimen/padding_button"
             android:textColor="@color/green"
             android:text="Login"/>
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:id="@+id/registration_token_text"
+            android:textColor="@color/white"
+            android:layout_marginTop="@dimen/padding_top"
+            android:padding="@dimen/padding_button"
+            android:text="No account yet? Sign up"
+            android:gravity="center"
+            android:textSize="16dip"/>
+
     </LinearLayout>
 
 </RelativeLayout>
